requirements.txt
Flask==2.0.1
Flask-SQLAlchemy==2.5.1
Flask-Migrate==3.1.0
Flask-Cors==3.0.10
Flask-JWT-Extended==4.3.1
Flask-RESTful==0.3.9
Flask-Mail==0.9.1
celery==5.3.4
redis==4.5.4
SQLAlchemy==1.4.23
marshmallow==3.14.0
python-dotenv==0.19.0
pytest==6.2.5
gunicorn==20.1.0
bcrypt==3.2.0
email-validator==1.1.3
Werkzeug==2.0.1
itsdangerous==2.0.1
Jinja2==3.0.2
MarkupSafe==2.0.1
alembic==1.7.4
requests==2.31.0
python-dateutil==2.8.2 